requirements.txt
# Core Dependencies
gradio>=4.44.0
httpx>=0.27.0
python-telegram-bot>=21.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Video Processing
ffmpeg-python>=0.2.0
pillow>=10.0.0

# Encryption (for Douyin API)
gmssl>=3.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# HTTP
requests>=2.31.0
aiohttp>=3.9.0

# HuggingFace
huggingface_hub>=0.23.0
datasets>=2.16.0
pandas>=2.0.0
